Fix column generation logic to be deterministic
There was a serious bug where on strip creation (multiple ticket creation in succession one after another), bingo rules were maintained but the first there were the issue due to the randomness that some columns were selected to be allocated more often than others, causing the last iteration to fail because the non selected columns needed more than 3 numbers to be allocated with.

This commit fix this issue by enhancing the business logic of the column selection, to select the column to allocate evenly using a pool of allocated numbers together with the previously allocated numbers.