Enhance service to prioritise underpopulated columns
This change was needed because there was a deterministic bug due to randomness selection of the columns, which caused earlier iterations to randomly not select any numbers to allocate to certain columns and thus towards the last iterations, remaining numbers to exist for allocation.

This was a breaking bug, and the solution was to introduce a mechanism that prioritize these under-allocated columns based on the #iteration.